+++ src/TankMonster.cpp
@@ -5,7 +5,7 @@
 
 TankMonster::TankMonster(
     QSvgRenderer* renderer,
-    short type,
+    size_t type,
     QGraphicsSvgItem* parent)
   : Monster(renderer, type, parent) {
   this->initializeMonster();
